% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfan-package.R
\docType{package}
\name{ggfan}
\alias{ggfan}
\alias{ggfan-package}
\title{ggfan: An implementation of Fan plots in ggplot2.}
\description{
For reference see \href{https://www.bankofengland.co.uk/quarterly-bulletin/1998/q1/the-inflation-report-projections-understanding-the-fan-chart}{Britton, E.; Fisher, P. & J. Whitley (1998) The Inflation Report Projections: Understanding the Fan Chart}.
}
